feat: 优化DriveSnapSeq自增操作 #3999

+++ src/backend/driver/boot-driver/src/main/kotlin/com/tencent/bkrepo/fs/server/config/properties/drive/DriveProperties.kt
@@ -4,5 +4,5 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties("drive")
 data class DriveProperties(
-    var listCountLimit: Long = 100000L
+    var listCountLimit: Int = 100000
 )
